feat(task-20): Error handling when there are no RSS articles available to generate a podcast

+++ src/rss-feeds-to-podcast/modules/config/schemas/rss.schema.ts
@@ -8,7 +8,7 @@ export const RssFeedSchema = z.object({
 
 export const RssConfigSchema = z.object({
   feeds: z.array(RssFeedSchema).min(1, 'At least one RSS feed is required'),
-  maxAgeHours: z.number().int().positive().optional(),
+  maxAgeHours: z.number().positive().optional(),
 });
 
 export type RssConfig = z.infer<typeof RssConfigSchema>;
